+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 - Expose status of wallet connections
 - Show banner for login alert
+- Show banner for system messages
 
 ## [2.0.7] - 2024-06-11
 ### Changed: